rename, remove: 중복 파일 삭제 및 파일 경로 삭제 - 기존 컨벤션에 맞지않는 패키지 경로 수정
