+++ src/main/java/com/bugflix/weblog/like/domain/Like.java
@@ -1,11 +1,9 @@
-package com.bugflix.weblog.like;
+package com.bugflix.weblog.like.domain;
 
 import jakarta.persistence.*;
-import jakarta.transaction.Transactional;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
 @Entity(name = "like_tb")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
