+++ src/main/java/com/bugflix/weblog/like/repository/LikeRepository.java
@@ -1,14 +1,11 @@
-package com.bugflix.weblog.like;
+package com.bugflix.weblog.like.repository;
 
+import com.bugflix.weblog.like.domain.Like;
+import com.bugflix.weblog.like.domain.LikeKey;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
 
 public interface LikeRepository extends JpaRepository<Like, LikeKey> {
     Like findLikeById_PostIdAndId_UserId(Long postId, Long userId);
-
-    List<Like> findById_PostId(Long postId);
 
     boolean existsLikeById_PostIdAndId_UserId(Long postId, Long userId);
