+++ src/main/java/com/bugflix/weblog/like/service/LikeServiceImpl.java
@@ -1,7 +1,8 @@
-package com.bugflix.weblog.like;
+package com.bugflix.weblog.like.service;
 
-import com.bugflix.weblog.like.dto.LikeStatus;
-import com.bugflix.weblog.user.User;
+import com.bugflix.weblog.like.domain.Like;
+import com.bugflix.weblog.like.dto.LikeStatusResponse;
+import com.bugflix.weblog.like.repository.LikeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -24,7 +25,7 @@ public class LikeServiceImpl {
      * Parameter :
      * - Long postId
      * return :
-     * - LikeStatus
+     * - LikeStatusResponse
      * - likeCount : 좋아요 갯수
      * - isLiked : 본인의 좋아요 상태
      * <p>
@@ -32,10 +33,10 @@ public class LikeServiceImpl {
      * - 이미 Post 에 Like 표시를 한 경우 : DB 에서 Like Entity 삭제 && isLiked == false && likeCount -= 1
      * - 그 반대의 경우 : DB 에 Like Entity 추가 && isLiked == true && likeCount += 1
      * <p>
-     * - isLiked 와 likeCount 를 LikeStatus 로 encapsulation 하여 반환
+     * - isLiked 와 likeCount 를 LikeStatusResponse 로 encapsulation 하여 반환
      */
-    public LikeStatus changeLikeStatus(Long postId) {
-        LikeStatus likeStatus = new LikeStatus();
+    public LikeStatusResponse changeLikeStatus(Long postId) {
+        LikeStatusResponse likeStatus = new LikeStatusResponse();
 
         Long userId = 1L;  // Todo User 정보 받아오기
