+++ src/main/java/com/bugflix/weblog/page/repository/PageRepository.java
@@ -1,9 +1,10 @@
-package com.bugflix.weblog.page;
+package com.bugflix.weblog.page.repository;
 
+import com.bugflix.weblog.page.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 public interface PageRepository extends JpaRepository<Page, Long> {
-    public Optional<Page> findPageByUrl(String url);
+    Optional<Page> findPageByUrl(String url);
 }
