+++ src/main/java/com/bugflix/weblog/page/service/PageServiceImpl.java
@@ -1,5 +1,7 @@
-package com.bugflix.weblog.page;
+package com.bugflix.weblog.page.service;
 
+import com.bugflix.weblog.page.domain.Page;
+import com.bugflix.weblog.page.repository.PageRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
