+++ src/main/java/com/bugflix/weblog/post/repository/PostRepository.java
@@ -1,5 +1,6 @@
-package com.bugflix.weblog.post;
+package com.bugflix.weblog.post.repository;
 
+import com.bugflix.weblog.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -9,7 +10,5 @@ public interface PostRepository extends JpaRepository<Post, Long> {
     public List<Post> findByPageUrl(String url);
 
     public List<Post> findByPageUrlAndUserUserId(String url, Long userId);
-
-    public void deleteByPostId(Long postId);
 
 }
