--- src/main/java/com/bugflix/weblog/profile/dto/ProfileRequest.java
@@ -1,8 +1,0 @@
-package com.bugflix.weblog.profile.dto;
-
-import lombok.Data;
-
-@Data
-public class ProfileRequest {
-
-}
