--- src/main/java/com/bugflix/weblog/tag/dto/TagResponse.java
@@ -1,4 +1,0 @@
-package com.bugflix.weblog.tag.dto;
-
-public class TagResponse {
-}
