+++ src/main/java/com/bugflix/weblog/tag/repository/TagRepository.java
@@ -1,13 +1,11 @@
-package com.bugflix.weblog.tag;
+package com.bugflix.weblog.tag.repository;
 
+import com.bugflix.weblog.tag.domain.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface TagRepository extends JpaRepository<Tag, Long> {
-
-    public Optional<Tag> findTagByTagContent(String tagContent);
 
     public List<Tag> findTagsByPostPostId(Long postId);
