+++ src/main/java/com/bugflix/weblog/tag/service/TagServiceImpl.java
@@ -1,6 +1,7 @@
-package com.bugflix.weblog.tag;
+package com.bugflix.weblog.tag.service;
 
-import com.bugflix.weblog.tag.dto.TagRequest;
+import com.bugflix.weblog.tag.domain.Tag;
+import com.bugflix.weblog.tag.repository.TagRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
