--- src/main/java/com/bugflix/weblog/user/UserController.java
@@ -1,4 +1,0 @@
-package com.bugflix.weblog.user;
-
-public class UserController {
-}
